+++ src/Clock.php
@@ -10,4 +10,15 @@ interface Clock extends ClockInterface
 
 	public function now(): DateTimeImmutable;
 
+	/**
+	 * Sleep for given amount of time.
+	 * Negative and zero sleep times are ignored.
+	 * If no arguments are given, warning is triggered.
+	 */
+	public function sleep(
+		int $seconds = 0,
+		int $milliseconds = 0,
+		int $microseconds = 0
+	): void;
+
 }